refactor: again moving things around
