+++ libs/wallet-react/src/components/connect-dialog/connect-dialog/index.tsx
@@ -1,0 +1,1 @@
+export * from './connect-dialog';
