+++ libs/wallet-react/src/components/connect-dialog/connection-options/connection-options.tsx
@@ -1,102 +1,13 @@
 import { useState } from 'react';
-import {
-  type Connector,
-  ConnectorErrors,
-  type ConnectorType,
-} from '@vegaprotocol/wallet';
-import {
-  Accordion,
-  AccordionItem,
-  DialogTitle,
-} from '@vegaprotocol/ui-toolkit';
+import { ConnectorErrors, type ConnectorType } from '@vegaprotocol/wallet';
+import { DialogTitle } from '@vegaprotocol/ui-toolkit';
 import { useT } from '../../../hooks/use-t';
 import { useWallet } from '../../../hooks/use-wallet';
 import { useConnect } from '../../../hooks/use-connect';
 import { Links } from '../../../constants';
 import { Trans } from 'react-i18next';
-import { ConnectionOptionRecord } from './custom-connection-options';
-import { ConnectionOptionDefault } from './connection-option-default';
-import { ConnectionOptionDeemphasizedDefault } from './connection-option-deemphasized-default';
-
-const ProminentConnectors = ({
-  onConnect,
-  setIsInstalling,
-  connectors,
-}: {
-  onConnect: (id: ConnectorType) => void;
-  setIsInstalling: (isInstalling: boolean) => void;
-  connectors: Connector[];
-}) => {
-  return (
-    <div className="grid grid-cols-1 sm:grid-cols-2 gap-1 -mx-2">
-      {connectors
-        .filter((c) => c.prominent)
-        .map((c) => {
-          const ConnectionOption = ConnectionOptionRecord[c.id];
-          const props = {
-            id: c.id,
-            name: c.name,
-            description: c.description,
-            showDescription: false,
-            onClick: () => onConnect(c.id),
-            onInstall: () => setIsInstalling(true),
-          };
-
-          if (ConnectionOption) {
-            return (
-              <li key={c.id}>
-                <ConnectionOption {...props} />
-              </li>
-            );
-          }
-
-          return (
-            <li key={c.id}>
-              <ConnectionOptionDefault {...props} />
-            </li>
-          );
-        })}
-    </div>
-  );
-};
-
-export const AdvancedConnectionOptions = ({
-  onConnect,
-  setIsInstalling,
-  connectors,
-}: {
-  onConnect: (id: ConnectorType) => void;
-  setIsInstalling: (isInstalling: boolean) => void;
-  connectors: Connector[];
-}) => {
-  const t = useT();
-
-  return (
-    <Accordion>
-      <AccordionItem
-        itemId="current-fees"
-        title={t('Advanced connection options')}
-        content={connectors
-          .filter((c) => !c.prominent)
-          .map((c) => {
-            const props = {
-              id: c.id,
-              name: c.name,
-              description: c.description,
-              onClick: () => onConnect(c.id),
-              onInstall: () => setIsInstalling(true),
-            };
-
-            return (
-              <li key={c.id}>
-                <ConnectionOptionDeemphasizedDefault {...props} />
-              </li>
-            );
-          })}
-      />
-    </Accordion>
-  );
-};
+import { ProminentConnectorsList } from './prominent-connectors-list';
+import { AdvancedConnectionOptionsList } from './advanced-connection-options-list';
 
 export const ConnectionOptions = ({
   onConnect,
@@ -126,10 +37,10 @@ export const ConnectionOptions = ({
       ) : (
         <>
           <ul className="w-full" data-testid="connectors-list">
-            <ProminentConnectors
+            <ProminentConnectorsList
               {...{ onConnect, setIsInstalling, connectors }}
             />
-            <AdvancedConnectionOptions
+            <AdvancedConnectionOptionsList
               {...{ onConnect, setIsInstalling, connectors }}
             />
           </ul>
